% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qq}
\alias{qq}
\title{Creates a Q-Q plot}
\usage{
qq(pvector, ...)
}
\arguments{
\item{pvector}{A numeric vector of p-values.}

\item{...}{Other arguments passed to \code{plot()}}
}
\value{
A Q-Q plot.
}
\description{
Creates a quantile-quantile plot from p-values from a GWAS study.
}
\examples{
qq(gwasResults$P)
}
\keyword{qq}
\keyword{qqplot}
\keyword{visualization}

