% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_sample.R
\name{qq_sample}
\alias{qq_sample}
\title{Creates a Q-Q plot}
\usage{
qq_sample(pvector, p_thresh = 1e-05, n_sample = Inf, ...)
}
\arguments{
\item{pvector}{A numeric vector of p-values.}

\item{p_thresh}{P-value threshold for sampling. P-values > \code{p_thresh} will be sampled.}

\item{n_sample}{Integer indicating number of sub-threshold points to plot}

\item{...}{Other arguments passed to \code{plot()}}
}
\value{
A Q-Q plot.
}
\description{
Creates a quantile-quantile plot from p-values from a GWAS study. 
Sub threshold points are sampled to reduce plotting time. Currently uses reservior sampling
}
\examples{
qq(gwasResults$P)

}
\keyword{qq}
\keyword{qqplot}
\keyword{visualization}
