% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqman-package.R
\docType{package}
\name{qqman-package}
\alias{qqman}
\alias{qqman-package}
\title{qqman: Q-Q and Manhattan Plots for GWAS Data}
\description{
Create Q-Q and manhattan plots for GWAS data from PLINK results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stephenturner/qqman}
  \item Report bugs at \url{https://github.com/stephenturner/qqman/issues}
}

}
\keyword{internal}
