% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{qqman}
\alias{qqman}
\alias{qqman-package}
\title{Create Q-Q and manhattan plots for GWAS data.}
\description{
A package for creating Q-Q and manhattan plots for GWAS
data. See the package vignette for details: \cr\cr
\code{vignette("qqman")}
}
\author{
Stephen Turner <\url{http://stephenturner.us}>
}

